chore: add react-native.config.js for linking
Previously if we had to link something customly we put it in the
package.json and with the help of  command we linked the things
we wanted to by running react-native link. but after the release of version 0.60
we have to create a file at root directory named  and give
the path of assets in there. then we run the react-native link command